Add files via upload
PDF was split into 3 due to 25MB file limit. Embedded youtube links may not work. THe HTML raw file is available in the same repository and has working embedded links that can redirect to Youtube. Formatting also overlaps some of the text; thus refer to the HTML source for full text if text cannot be extracted.