Add files via upload
"Although Morgan Ames does good research work that fosters credibility in the reader, her depiction of the
OLPC project may be perceived as excessively critical, which could lead readers towards scepticism of
educational technology initiatives in general and constructionism in particular. While it is widely acknowledged
that the OLPC initiative failed to achieve its intended goals, it is important to note that the underlying
principles of constructionism, and the aspiration to provide equal access to education for all children, are
commendable. However, the book fails to capture the positive nuances of these principles and aspirations."