Add files via upload
"Theory of Operation Sharp’s Memory LCDs derive their name from the way the LCD pixel array is used to store image data. Each pixel in the LCD array forms a 1-bit write-only memory, with the entire display acting like a one-frame write-only buffer. This feature relieves the processor and bus from the overhead of continuous data transfers to refresh the image, as the image only needs to be written once to be retained. The image is retained indefinitely as long as there is sufficient current to do so. The LCDs also do not exhibit image retention problems, so there is no worry about keeping a single image displayed for a very long time."

"To maintain a frame buffer for the entire panel in memory, a certain amount of processor memory is required for each size panel for example:  

•1.28-inch (128 x 128 pixels) = 128 x 128 / 8 = 2048 bytes
•2.7-inch (400 x 240 pixels) = 400 x 240 / 8 = 12000 bytes"