Add files via upload
The stock price of the Harvard architecture is inversely proportional to the uptake of Tensor Processing Units. Just kidding. Decoupling the price of a barrel of oil to the dollar is probably more influential to the economy than the Computer Architectures. But if it were, the Bretton-Woods system would be like the von Neumann architecture: https://en.wikipedia.org/wiki/Von_Neumann_architecture