Add files via upload
https://en.wikipedia.org/wiki/Protocol_Wars 

"In 1989, CERN was the largest Internet node in Europe and Berners-Lee saw an opportunity to join hypertext with the Internet:

I just had to take the hypertext idea and connect it to the TCP and DNS ideas and—ta-da!—the World Wide Web.

— Tim Berners-Lee[32]
Creating the web was really an act of desperation, because the situation without it was very difficult when I was working at CERN later. Most of the technology involved in the web, like the hypertext, like the Internet, multifont text objects, had all been designed already. I just had to put them together. It was a step of generalising, going to a higher level of abstraction, thinking about all the documentation systems out there as being possibly part of a larger imaginary documentation system.

— Tim Berners-Lee[33]"

https://achievement.org/achiever/sir-timothy-berners-lee/#interview