Add files via upload
"Bose and Saha both went on to become physics instructors at the University of Calcutta. But due to Bengal’s isolation and the effects of
the First World War, they found it hard to follow the most recent developments in Europe. One of the few periodicals that was regularly
available in the Presidency library was Philosophical Magazine, in which Bose and Saha read one of Niels Bohr’s seminal papers on atomic
structure, published in 1913 (Phil Mag. 26 1).

In Calcutta, they were also fortunate to befriend Paul Johannes Brühl, a visiting botanist from Germany, who had brought with him books
and journals on thermodynamics, quantum theory, relativity and other popular physics topics. In 1919, after Einstein had shot to fame
following the apparent confirmation of general relativity, Bose and Saha managed to get copies of the basic papers in German and French.
Bose was fluent in both languages, as well as English, and so he and Saha translated and published the papers in book form as The
Principle of Relativity (University of Calcutta, 1920). It was the first English-language collection of papers on the topic from Einstein and
others" https://physicsworld.com/a/when-bose-wrote-to-einstein-the-power-of-diverse-thinking/

https://archive.org/details/principleofrelat00eins