Add files via upload
https://www.theguardian.com/books/2018/sep/20/the-coddling-of-the-american-mind-review

https://indico.cern.ch/event/1331906/contributions/5606846/attachments/2743709/4773509/TCPIP_CS40.pdf

https://semiengineering.com/a-power-first-approach/ 
"Even today, when I interview people and ask if power is a primary optimization consideration, the answer is typically, “We have to meet performance metrics first and then we can worry about reducing power.” Over the next 10 years, I expect that mentality will have changed to a “power first” approach. The real question is, within a defined power budget, how much performance on a given task can I attain? You don’t get to that answer by looking at performance first."

"But could we now be seeing the genesis of what will be the new era? New abstractions, flows, and tools where the primary focus is on power and energy? A new generation of design that will be able to maximize performance per watt, not through optimization, but through design?"

This quote is in reference to my forum post where I suggest (strongly) that 5mW be a TDP limit. The unitalicized quotes at the in the RWT forum post are from page four of the NTV article from 2012. Keep in mind, this site is frequented by tech luminaries and is very much like a lions den of sorts. 

https://www.realworldtech.com/near-threshold-voltage/4/ 

"Weak-link problems are problems where the overall quality depends on how good the worst stuff is. 
 fix weak-link problems by making the weakest links stronger, or by eliminating them entirely."
https://www.experimental-history.com/p/science-is-a-strong-link-problem

NTV is like a weak link problem. Intel cancelled their Quark processor because it didn't have an immediate market outside Galileo boards and hobbyists that tinkered with it like the Raspberry Pi.

Instead it could have strengthened the minimum performance of it by adding sufficient RAM.

If the experts were wrong, then they need to rethink the the markets that they selected their markets foes not HPC and video processors but general computing 

And that is why this philosopher may be right here:
"Here’s how Feyerabend put it in Against Method (1975, emphasis his):

The thesis is: the events, procedures and results that constitute the sciences have no common structure; there are no elements that occur in every scientific investigation but are missing elsewhere. [...] Successful research does not obey general standards; it relies now on one trick, now on another [...] This liberal practice, I repeat, is not just a fact of the history of science. It is both reasonable and absolutely necessary for the growth of knowledge."

https://www.experimental-history.com/p/the-anarchist-and-the-hockey-stick