Add files via upload
https://github.com/EI2030/Low-power-E-Paper-OS/commit/f101bdc283f1473c28df1b58f4639716a2d1f6a7#commitcomment-156460484 

"Telesatellites and mobile networks
This year marks 40 years since the Storting first allocated money for Norwegian telecommunications research. For the first 13 years, Halvor Bothner-Bye was one of the key professionals who led the Norwegian Telecommunication Authority's Research Institute. Bothner-By left TF in 1981, but played and still plays important roles in the Norwegian telecommunications industry, both as an entrepreneur and manager of his own company and in management positions such as chairman of the board of directors at Høykom. When we accepted an offer from Halvor Bothner-By for an article about 40 years of telecommunications research in Norway, we ordered 7,000 characters. But the writing engaged the telecommunications veteran so much that he delivered nearly 40,000 characters, and was told that he could not cut any more from the text. We therefore bring excerpts of the history here in Teleavisen on paper and publish everything in its entirety on teleavisen.no. After a discussion of the organization and working methods at TF, Bother-By tackles areas where the institute made innovative efforts. Here are some of them:" (2007)

https://www.digi.no/artikler/telesatellitter-og-mobilnett/121362

https://en.wikipedia.org/wiki/Halvor_Bothner-By

https://onlinelibrary.wiley.com/doi/abs/10.1002/ett.4460030513

https://www.eecis.udel.edu/~amer/PEL/survey/survey_tables.pdf

https://web.archive.org/web/20190228065633/http://old.open-root.eu/la-doc/histoire-des-reseaux/le-datagramme/

https://blog.eero.com/how-did-people-communicate-before-the-internet/

https://www.gulf-insider.com/wp-content/uploads/2024/09/APRIL-2022.pdf (p.17/80)